Add support for Neutron RBAC sharing of networks
Networks that will be shared by Neutron RBAC are assumed to be
owned by the admin project and private to that project.
Such networks can be shared with a list of projects either in
"access_as_external" or "access_as_shared" mode.

The RBAC assignments cannot be modified once they have been created.